intellitask/requirements.txt
asgiref==3.7.2
Django==5.0.3
django-debug-toolbar==4.3.0
django-environ==0.11.2
djangorestframework==3.14.0
environ==1.0
pillow==10.2.0
psycopg2-binary==2.9.9
pytz==2024.1
sqlparse==0.4.4
tzdata==2024.1
